[frontend] remove web3 functions file
